requirements.txt
pg8000==1.31.1
SQLAlchemy==2.0.29
google-cloud-aiplatform==1.46.0
google-cloud-alloydb-connector==1.0.0
pandas
google-auth
